Move DateSelection in Editor to dependent component (mac)
